Enable deprecation silencing in babel workers
